Import from http://sys.cs.uos.de/bonnmotion/src/bonnmotion-3.0.0.zip

+++ validate/AppValidationDispatcher.py
@@ -1,18 +1,24 @@
-#    A validation script for Bonnmotion (http://net.cs.uni-bonn.de/wg/cs/applications/bonnmotion/)
-#    Copyright (C) 2011 University of Bonn
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -*- coding: utf-8 -*-
+################################################################################
+## A validation script for                                                    ##
+## BonnMotion - a mobility scenario generation and analysis tool              ##
+## Copyright (C) 2002-2012 University of Bonn                                 ##
+## Copyright (C) 2012-2015 University of Osnabrueck                           ##
+##                                                                            ##
+## This program is free software; you can redistribute it and/or modify       ##
+## it under the terms of the GNU General Public License as published by       ##
+## the Free Software Foundation; either version 2 of the License, or          ##
+## (at your option) any later version.                                        ##
+##                                                                            ##
+## This program is distributed in the hope that it will be useful,            ##
+## but WITHOUT ANY WARRANTY; without even the implied warranty of             ##
+## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the              ##
+## GNU General Public License for more details.                               ##
+##                                                                            ##
+## You should have received a copy of the GNU General Public License          ##
+## along with this program; if not, write to the Free Software                ##
+## Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA  ##
+################################################################################
 
 from Config import Config
 from Common import log, runBonnmotionApp, Hashes
@@ -59,9 +65,10 @@ class AppValidationDispatcher(object):
             for i in t.Seq:
                 os.remove(os.path.join(Config().readConfigEntry('bonnmotionvalidatepath'), Config().readConfigEntry('tempoutputmovementsfile').replace('INDEX', str(i.no))))
                 os.remove(os.path.join(Config().readConfigEntry('bonnmotionvalidatepath'), Config().readConfigEntry('tempoutputparamsfile').replace('INDEX', str(i.no))))
-                os.remove(Config().readConfigEntry('tempoutputparamsfile').replace('INDEX', str(i.no)))   
+                
                 for x in i.ordering.split(','): 
-                    os.remove(os.path.join(Config().readConfigEntry('bonnmotionvalidatepath'), Config().readConfigEntry('tempoutputname') + str(i.no)) + '.' + x)          
+                    if x != 'NULL':
+                        os.remove(os.path.join(Config().readConfigEntry('bonnmotionvalidatepath'), Config().readConfigEntry('tempoutputname') + str(i.no)) + '.' + x)          
     
     #######################################
     #Value class. To encapsulate the parameters an AppValidationThread needs.
@@ -96,16 +103,18 @@ class AppValidationDispatcher(object):
                 spliting = i.ordering.split(',')
                 if len(spliting) > 1:                       #the app creates more than 1 file
                     for ext in spliting: 
-                        f = open(outputfilename + '.' + ext)
-                        content = content + f.read()        #concatenate content of all files created
-                        f.close()
+                        if i.ordering != 'NULL':
+                            f = open(outputfilename + '.' + ext)
+                            content = content + f.read()        #concatenate content of all files created
+                            f.close()
                 else:                                       #the app creates only one file
-                    f = open(outputfilename + '.' + i.ordering)
-                    content = f.read()
-                    f.close()
+                    if i.ordering != 'NULL':
+                        f = open(outputfilename + '.' + i.ordering)
+                        content = f.read()
+                        f.close()
     
                 if (i.md5 <> Hashes().md5(content) or i.sha1 <> Hashes().sha1(content)):
-                    f2 = open(Config().readConfigEntry('tempoutputparamsfile').replace('INDEX', str(i.no)), 'r')
+                    f2 = open(os.path.join(Config().readConfigEntry('bonnmotionvalidatepath'), Config().readConfigEntry('tempoutputparamsfile').replace('INDEX', str(i.no))), 'r')
                     if i.appparameters != '': 
                         self.ReturnValue.append('parameters of bonnmotion:\n' + f2.read() + '\nappname: ' + i.identifier + '\nparameters of the app:\n' + i.appparameters)
                     else:
